requirements.txt
fastapi
uvicorn
transformers
beautifulsoup4
newspaper3k
torch
requests
python-dotenv
lxml_html_clean
readability-lxml
lxml